+++ druid/src/core.rs
@@ -678,7 +678,7 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                     }
                 }
             },
-            Event::WindowConnected => true,
+            Event::WindowConnected | Event::WindowCloseRequested => true,
             Event::WindowDisconnected => {
                 for (window_id, _) in &self.state.sub_window_hosts {
                     ctx.submit_command(CLOSE_WINDOW.to(*window_id))